fix(fonts): scan macOS font directories directly instead of fc-list
macOS does not ship fontconfig by default, so GetSystemFonts() calling
`fc-list` (fonts_unix.go, previously shared with Linux) silently failed
on stock Macs. The frontend caught the error and fell back to a tiny
hardcoded 6-font list, so any real system or user-installed font
(including one just installed to ~/Library/Fonts) never appeared in
the font picker even though every other app on the machine could see
and use it fine.

- fonts_darwin.go (new): walk the standard macOS font directories
  (~/Library/Fonts, /Library/Fonts, /System/Library/Fonts[/Supplemental])
  directly and parse each .ttf/.otf/.ttc file, no external binary
  dependency.
- fonts_ttf.go (new, windows || darwin): extracted the shared TTF/OTF/TTC
  binary parser so Windows and macOS reuse identical font-parsing logic
  instead of duplicating it. Along the way, fixed two real bugs in the
  parser surfaced by testing against this Mac's actual font set:
    1. TTC (TrueType Collection) offsets inside a sub-font's table
       directory are absolute offsets into the whole file, not relative
       to that sub-font's table directory. The original code re-sliced
       `data` to "enter" a sub-font, which shifted every later offset
       read by that amount and decoded random glyph/binary data as the
       name table — producing mangled family name garbage for most
       .ttc fonts (Avenir, Baskerville, Charter, Muna, etc., none of
       which are actually monospace).
    2. Monospace detection relied solely on the post table's
       isFixedPitch flag, which is unreliable in practice — e.g. macOS's
       own Monaco.ttf renders as monospace but ships with this flag
       cleared to 0, so Monaco/Andale Mono silently disappeared from the
       list. Now falls back to measuring actual hmtx glyph advance
       widths (>=85% sharing one width) when the flag is unset.
- fonts_unix.go: build tag narrowed to `!windows && !darwin`, so the
  fc-list path now only applies to Linux, where fontconfig is a safe
  default assumption.

Verified against this machine's real /System/Library/Fonts,
/System/Library/Fonts/Supplemental, and ~/Library/Fonts: output now
correctly includes Monaco, Andale Mono, Courier New, and a
just-installed DejaVu Sans Mono, with no garbled family names and no
duplicates.
